requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg[binary,pool]>=3.1
sqlalchemy==2.0.30
pymongo==4.6.2
pandas==2.2.2
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0